refactor: move to reusable component
