+++ src/css/query_editor.css
@@ -28,7 +28,7 @@ input[type='checkbox'] {
   margin: 4px;
 }
 
-.custom-query {
+.custom-query input {
   font-family: monospace;
 }
